Transition to ULocale.
This CL switches over the Locale implementation to use ICU4J's
ULocale instead of the ICU4C implementation. The methods that
used to call through to ICU4C were:

getDisplayLanguage, getDisplayCountry, getDisplayVariant,
and getDisplayScript for the display methods;

getISO3Country, getIso3Language for ISO codes;

getAvailableLocales to return an array of included system
Locales.

The toLanguageTag now delegates to ULocale's implementation
but covers up special cases relating to the POSIX variant
tag, deprecated language tag mappings, and undefined language
behaviour.

This change is stable (i.e. has no dependency cycles) with
ULocale because Locale does not depend on ULocale for static
initialization or construction of a Locale object. The only
time a ULocale object is created is when one of the display
methods, or toLanguageTag is called.

Additionally this CL switches over the methods to get available
locales for Calendar, Collator, DateFormat, DecimalFormat to
use the ICU4J equivalent. These return plain Locale objects
so no effort is needed to work with ULocales here. An effort
has been made to compare the list of returned locales from
each of these methods on the ICU4C and ICU4J implementation
and the language tags of the locales seem to be the same for
both.

Use new language codes for Hebrew, Yiddish, and Indonesian in
Locale.toLanguageTag() only.

Also updates POSIX language tag to use the proper extension
-u-va-posix as specified by BCP47.

toLanguageTag is now more compliant to the BCP47 spec, allowing
language tags which are 2*3ALPHA or 4ALPHA or 5*8ALPHA. Any
other language tag will return "und" when passed through
the toLanguageTag() method.

Change-Id: I9e8e50d5283dca0eee89bee9e21dfc2e3315a804