Transition to ULocale.
This CL switches over the Locale implementation to use ICU4J's
ULocale instead of the ICU4C implementation. The methods that
used to call through to ICU4C were:

getDisplayLanguage, getDisplayCountry, getDisplayVariant,
and getDisplayScript for the display methods;

getISO3Country, getIso3Language for ISO codes;

getAvailableLocales to return an array of included system
Locales.

The toLanguageTag now delegates to ULocale's implementation
but covers up special cases relating to the POSIX variant
tag, deprecated language tag mappings, and undefined language
bahaviour.

This change is stable (i.e. has no dependency cycles) with
ULocale because Locale does not depend on ULocale for static
initialization or construction of a Locale object. The only
time a ULocale object is created is when one of the display
methods, or toLanguageTag is called.

Change-Id: I9e8e50d5283dca0eee89bee9e21dfc2e3315a804