Provide the ability to pause/resume CloseGuard warnings
Some APIs provide no mechanism for explicit clean up.

Example: ClassLoader, URLConnection.

Legitimate use of these APIs from apps and within framework
code can lead to CloseGuard warnings that developers can do
nothing about. This change provides a mechanism to temporarily
disable CloseGuard for a specific thread to deal with these
cases.

Included is an example usage.

The implementation does not use ThreadLocal because the
expense of a ThreadLocal miss is relatively high and pausing
is expected to be a rare case.

Bug: 23990903
Change-Id: Ia57d6af3730f9c2d48b6cf1158c626c1dbbfd168