Posix: gracefully handle java/native code differences
Some posix functions take a size_t argument, but java doesn't have
an unsigned integer. In java, these functions take a signed value,
but the behavior of the functions when handling a negative value
aren't defined.

Throw an IllegalArgumentException when handling certain negative
java values, rather than having the value be interpreted as a
positive number when passed to the JNI layer.

Longer term, it would be best to enable
  LOCAL_SANITIZE := integer
which would detect improper integer conversions.

Change-Id: I78a8b99cb6d5a04901e3b16639c847279be30642