Remove CloseGuard from dalvik.system.DexFile
Also made DexFile.close a NOP since it may be called when the DexFile
is still in use. Closing the dex file is now all handled by the
finalizer. The reason for not enabling close() is the double close
issue. If someone calls close but the dex file is still in use, then
we cannot delete the dex file without crashing the runtime. And we
do not have any way to signal that the close operation was ignored
with the current API.

Bug: 22720414

Change-Id: I8f30549b73e48a600595ef80e5d1e67628af7ef0