Switch from jarjar to generated source for ICU4J
Future changes will include ICU4J classes as public APIs
(still beneath android.icu). To generate these docs source
code will be needed, as well as the ability to apply
changes to the javadoc comments programatically.

This change includes the source code generated using
tools in external/icu/tools/srcgen.

A consequence of this change is that libcore code can now
depend on icu in the android.icu package and not the
com.ibm.icu package as was the case previously. A happy
side effect is that the source code is available for use
during debugging.

Bug: 22023363
Change-Id: Icba5668fae53e2dc98ccc74e2f1ad1ca226e5feb