More efficient JCA crypto with read-only ByteBuffers.
The default implementations of JCA crypto SPI MessageDigestSpi,
SignatureSpi, CipherSpi, MacSpi methods which read input from
ByteBuffer are currently penalizing the use of read-only ByteBuffer.
This is because read-only ByteBuffer does not expose its contents as
a byte[] because byte[] is mutable. The penalty currently incurred
when using a read-only ByteBuffer as input is that a new byte[] of
length remaining() is allocated and then the contents of the
ByteBuffer are copied into that array.

Other parts of libcore work around the read-only input ByteBuffer
issue by obtaining the underlying backing array using hidden API
java.nio.NioUtils.unsafeArray(ByteBuffer). The assumption is that the
user of unsafeArray will be careful enough to not modify the returned
byte[].

This change makes the default implementations of MessageDigestSpi,
SignatureSpi, CipherSpi, MacSpi methods which read input from
ByteBuffer obtain the backing array of the ByteBuffer in the same way
as other performance-critical parts of libcore: using hidden API
NioUtils.unsafeArray. A danger is that third-party subclasses of
these four JCA SPIs may then see the resulting byte[] in their
update(byte[], ...) methods and might modify the byte[] thus modifying
the contents of the read-only ByteBuffer. However, given that these
update(byte[], ...) methods are not supposed to modify the input
byte[], it's an acceptable performance-risk trade-off.

Change-Id: I3e41a3647b00d1308d5470e4d4c019edeb0e98bb