Wrap the InputStream in a BufferedInputStream if necessary
The InputStream passed to the setInput(InputStream is, String)
method on KXmlParser is sometimes a BufferedInputStream but
mostly isn't (according to a brief analysis of the usages in
AOSP) that makes parsing less efficient. Rather than change all
the usages to explicitly wrap in BufferedInputStream this does
it within the method itself, but only if it has not already
been done.

Bug: b/20849543
Change-Id: I1d6da83448b0371c13b97c14be45ee51f7252687