Modify tests to make them less fragile
The tests are testing how Process class manages its
InputStreams and uses a couple of executable that will write to
either stdout or stderr. Unfortunately, the behaviour of those
executables (ls / mkdir) is affected by external factors such as
the current working directory and permissions of the running
process that can make the test fragile. This change switches to
using simpler processes whose behaviour in the cases being
tested is more closely defined.

Bug: 24900179
Change-Id: I702939d09295e749f5efd71105d91180999b969a