Delete unnecessary notify
Reference.enqueue calls ReferenceQueue.enqueue which already does an
notify on the internal lock object.

The notify on the reference queue was causing the test to be flaky
since the remove could finish before the child thread had called
rq.notify. This would then proceed to reset rq to a new reference
queue which was no longer synchronized causing rq.notify to throw a
IllegalMonitorStateException exception.

Bug: 26306978

Change-Id: I602524abdfd7a7c84688b8d4583ee83cacda83cb