Add some more tests for SimpleTimeZone
The pre-enso version of SimpleTimeZone doesn't support
specifying either WALL_TIME, STANDARD_TIME or UTC_TIME in its
most specific constructor. Using that constructor ends up with a
corrupted SimpleTimeZone that doesn't behave properly according
to the JavaDoc. Given that it works in Enso and has been in the
code since it was copied from harmony we do not have any plans
to fix it. However, adding additional tests will help protect us
from future breakages and provide us a base line against which
to make the changes in earlier releases if we find that we need
to do so.

Bug: 26459425
Change-Id: Ic9b94c7b33d56a55aa6502cc2c21b933f5b91117