Be consistent between Runtime.nativeLoad and ClassLoader.findLibrary.
ClassLoader.findLibrary looks at system and class loader's library
directories, but nativeLoad (which is used for doing the actual load
of the found file) is only giving the class loader's library
directories to the native linker. That meant that to resolve
dependencies, the linker was not looking at the system library, which
is inconsistent with ClassLoader.findLibrary.

Change-Id: I06e991cc0cda60dfa4e4c7072289f81ff8f0cd87