Fix NetworkInterface cannot discover down links
Second attempt in fixing the issue.

The OpenJDK implementation is incorrect since it does not return all
the interfaces on the machine, rather only interfaces with either IPv4
or IPv6 addresses associated with them. This fix restores the libcore
behavior, by reading the interfaces list from /sys/class/net to
populate the down interfaces.

Bug: 26238832
Change-Id: Ie59b938a78fe8ddfe87dbb0e2ede7bfcf0bee77a