Make test_clear_45877 more reliable and test more behavior
The test depends on the System.currentMillis() method and so its
behavior varies based on the time. It was failing consistently
but only when the current time was in the afternoon in Los
Angeles. This is a change in behavior from M but it looks as
though the old behavior was wrong, the new behavior is
consistent with OpenJDK and the test was only added as part of
a documentation update so doesn't look as though it was trying
to enforce the M behavior.

The test has been made more robust (uses a fixed time) and been
split into two tests, one of which tests the morning and one the
afternoon behavior.

Bug: 26022884
Change-Id: I15c8a9173a0819798bec3f11146fa7d5ea85600d