Make ZoneInfoTest.testReadSerialized work with CTS
CTS uses proper JUnit which sets the name of the TestCase
properly. Vogar uses its own version which does not set the name
properly. This inconsistency, along with an inadvertant
dependency of the serialized form of the ZoneInfo on the test
name led to the test working on Vogar but failing in CTS.

Bug: 26928688
Change-Id: Iba1e0acc6fcb40a49a966045c4e216bb96250bee