SSLSessionTest: compare equality of SSLSession by contents
If the SSLSession received from the cache was not the exact same object,
the #getId() tests would fail. Instead compare the SSLSession by its
contents.

This has cropped up now because we have ExtendedSSLSession in addition
to SSLSession. Conscrypt, in order to maintain backward compatibility
for unbundled builds, wraps its internal SSLSession implementation with
a delegating subclass of ExtendedSSLSession in order to implement newer
method calls in terms of its internal implementation without sacrificing
the ability to build against previous versions.

Bug: 27123298
Change-Id: Id6b446b146941532373ab7fcfc6b3ca464b3aabf