Reset seeds of SecureRandom objects after fork
This change will prevent reusing the same seed/state
of SecureRandomImpl object in java.util.UUID and
javax.crypto.JceSecurity after zygote fork. This scenario
would happen only when the SRI object is used before the fork
(these SecureRandomImpl instances are created
without explicit seed, and they generate one during their
first use), which is not happening atm, but we don't want
to depend on this.

Bug: 27273467
Change-Id: Ic5f3dcbb820b0f05666c41b22ca400af2b683137