BufferedInputStream: Rewrite skip() strategy.
On Android-M and below, calling skip() on a marked stream that
was below the mark limit would fill the buffer at least once.

OpenJdk did not offer that guarantee, and unsurprisingly, several
Android apps have grown to rely on it. The most common abuse is that
they don't check the return value of skip().

This change also makes sure the mark is invalidated if we skip beyond
the mark limit.

Despite this change, we still have a tiny difference in behaviour: If
we call skip() on a marked stream when the buffer is empty and
(n > buffer.length), we will skip |buffer.length| bytes and not
|2 * buffer.length| bytes. Fixing this is tricky and I'd rather not go
there unless we need to.

bug: 27191898
Change-Id: I1b02b5f3809ce4337f4d56f33c645934c554b413