OldManifestTest : Don't depend on HashMap iteration order.
The switch to OpenJdk changed iteration order (due to to a changed
secondary hash) but this test somehow continued to work because the
initial HashMap size used in Attributes changed (from 4->16). Now that
we've changed OpenJdk to use the same secondary hash as Android M, we're
observing differing iteration orders.

bug: 28082128
Change-Id: I55f6a6e96e61970edb01d5e26274cd577ff32d25