Avoid additional Object allocations that are immediately unboxed
This should catch most of the remaining cases in libcore, although
I was more hesitant in tests; for example, TreeMapExtendTest does
a lot of unnecessary boxing but the benefit of optimising the test
doesn't seem worth the risk of potentially introducing bugs.

I haven't explicitly checked if any of this overlaps with narayan's
changes - if yes, then hopefully it'll merge cleanly.

Bug: 28289401
Change-Id: I911d2b5cf130f974427b4fe7e1916f1632e5f4fa