Made NET_SockaddrToInetAddress fail on a bad family.
NET_SockaddrToInetAddress was very lenient and
made assumption that if address was not AF_INET6
then it had to be AF_INET. Which is not true
when kernel returns address filled with random
values. Now, it should throw SocketException
in this case.

I've checked others users of NET_SockaddrToInetAddress,
and they should be fine with an exception coming from
it.

Bug: 28347599
Change-Id: I71e36c9ccb2b882f1f570c898716e88dceb7d3d2