Add a test for the behavior of Content-Type defaulting
When some HTTP methods are used, the Content-Type header
is defaulted to "application/x-www-form-urlencoded" iff
it is not explicitly set. The test also includes checks
for what happens when a body is present for methods that
don't permit it.

This test revealed a behavior change in OkHttp 2.7.5
for OPTIONS requests. Previous versions used in Android
releases would reject request a body, but now they are
permitted. This is in line with the spec:

https://tools.ietf.org/html/rfc7231#section-4.3.7

The test reflects the new behavior.

Bug: 30492501
Bug: 27908885
Change-Id: I716f032388d6ddd079e7d320004885122ec6c0d6