Tests that Processes with redirects expose null input/output streams.
The terms "null input stream" and "null output stream" are defined
in the toplevel ProcessBuilder documentation; therefore this CL adds
the testing logic to ProcessBuilderTest. Note that the guarantee
that Process.get{Input,Output,Error}Stream() is of such a nature is
made both in the ProcessBuilder and in the Process documentation.

Note that the documentation doesn't guarantee that the package private
classes ProcessBuilder.Null{Input,Output}Stream are used for this; it
is also not the case for redirectErrorStream(true). Therefore this
test does not check that they're instanceof Null*Stream. Instead, it
checks some slightly weaker properties that must be true for null
streams must always fulfill.

Tested:
  cts-tradefed run cts -c libcore.java.lang.ProcessBuilderTest

Bug: 27464570
Change-Id: I9ddb8e006f943548884566e7b3ac34b4b949b120