Tests that Processes with redirects expose null input/output streams.
The terms "null input stream" and "null output stream" are defined
in the toplevel ProcessBuilder documentation; therefore I have put
this testing logic into ProcessBuilderTest. Note that the guarantee
that Process.get{Input,Output,Error}Stream() is of such a nature is
made both in the ProcessBuilder and in the Process documentation.

I considered checking that the stream's class is specifically
ProcessBuilder.Null{Input,Output}Stream, but have refrained because
  - the contract doesn't specifically guarantee this
  - it's not actually true in the case of redirectErrorStream(true).

The test assumes that a stream that passes the test meets the
requirements for "null output/input stream", although of course
it can't know whether the stream is stateful and may fail the test
in other states.

Tested:
  cts-tradefed run cts -c libcore.java.lang.ProcessBuilderTest

Bug: 27464570
Change-Id: I9ddb8e006f943548884566e7b3ac34b4b949b120