Tests that Processes with redirected streams expose null input/output streams.
The terms "null input stream" and "null output stream" are defined
in the toplevel ProcessBuilder documentation, although the guarantee
that Process.get{Input,Output,Error}Stream() is of such a nature is
in the Process documentation.

Since that documentation refers specifically to ProcessBuilder as
opposed to Processes constructed otherwise, I'm not sure whether
this test should live in ProcessTest or ProcessBuilderTest. For now,
I've put it in ProcessTest, but I'm also happy to move it.

I've followed Process's lead in using tags such as
   <a href="ProcessBuilder.html#redirect-input">
in the documentation, even though I'm not sure if those are well
formed.

I considered the stricter check of stream.getClass().getName() being
ProcessBuilder.Null{Input,Output}Stream, but this stricter condition
doesn't currently hold for the redirectErrorStream(true) case.

My behavior checks don't cover all potential bugs. For example, I can't
think of a way to check that "close() does nothing".

Tested:
  cts-tradefed run cts -c libcore.java.lang.ProcessTest

Bug: 27464570
Change-Id: I9ddb8e006f943548884566e7b3ac34b4b949b120