prepare for okhttp upgrade: Fix URLConnectionTest.testConnectTimeouts
I fixed this test by dropping a dependency on Mockito which no
longer works with okhttp 2.7.5 for some reason. The new version
of the test should also work with the old code, so I'm changing
it ahead of the planned okhttp upgrade.

More detail:
In http://r.android.com/169087 this code changed to use spy() fake.
With okhttp Mockito 2.7.5, this test fails in mockito's
ConditionalStackTraceFilter.<init> because the
ClassPathLoader.getStackTraceCleanerProvider() returns null.
I suspect it's due to a bug in the order of static initializers
being run (the method has a TODO from 2013 to throw a  better
exception). I haven't figured out why it appears to happen only
after upgrading okhttp.

Since the dependency on Mockito isn't actually required for this
fake, I've implemented our own fake logic.

I've also tweaked the comment in URLConnectionTest which previously
claimed that the socketSpy was a mock (as opposed to a fake) and
that it 'wraps the actual socket', both of which I believe were
incorrect.

Tested:
make cts && cts-tradefed run cts -c libcore.java.net.URLConnectionTest

Bug: 28935102
Change-Id: I792eef20fd9744b9e9b7a5b4b32ddf078af3c433