Fix URLConnectionTest.testConnectTimeouts
This drops a dependency on Mockito which no longer works for some reason
with okhttp 2.7.5. The new version of the test works with the old okhttp
code, so this change tweaks it ahead of the planned okhttp upgrade.

More detail:
 - org.mockito.internal.creation.CglibMockMaker fails to be instantiated
   during ClassPathLoader.<clinit>. This results in a MockitoException
   attempted to be thrown but instead throws ExceptionInInitializerError
   because MockitoException.<init> in turn depends on
   ClassPathLoader.<clinit> (via getStackTraceCleanerProvider()) which
   has not yet completed.
 - Unlike the old spy() wrapper from http://r.android.com/169087
   the new fake object will not forward additional method calls to
   the real socket. This isn't ideal, but not critical for this test.

Tested:
make cts && cts-tradefed run cts -c libcore.java.net.URLConnectionTest

Bug: 29095428
Change-Id: I792eef20fd9744b9e9b7a5b4b32ddf078af3c433