Remove default ProxySelector socket tunneling
If ProxySelector#getDefault is not empty, openJdk
SocksSocketImpl (the default SocketImpl) will try to
use it to proxy the socket connection through it, even
if the socket was created without the Proxy instance.

This is a regression from marshmallow release, which
would only tunnel the socket if the Proxy was explicitly
set in constructor.

Bug: 29092095
Change-Id: I0f59e8198804adb5890062fcacab457165876f4a