Correct address fields in ByteBufferAsXBuffer classes
The value of the address field should be equals to the first element of
the ByteBufferAsXBuffer class, previously, it was pointing to the first
element of the underlying directbytebuffer object.

Also, which creating duplicates, slice and readonly variants of
DirectByteBuffer, the order was not preserved. Corrected the same as
well.

Bug: 28964300
Change-Id: I6dfc98ba408c524580198ea168923981b23a0bc2