Fix Preferences regression.
This change fixes reading filesystem-saved
java.util.prefs.Preferences.

Due to bug in XmlSupport, when reading the
filesystem-saved preferences, reader was
throwing an exception. Android
implementation of xml parser adds additional
nodes that were not Elements, this was
causing a cast exception (old code was expecting
only Elements).

Now, this exception is silently swallowed by the
FileSystemPreferences if it happens during Preferences
content initialization.... I've added additional
log statement to help with this.

Bug: 27645233
Change-Id: Ieebd8a6a1002111654005bfcc42009b01c89ea70