SerializationStressTest4 : Improve iteration tests.
In tests :
- test_writeObject_Collections_UnmodifiableCollection
- test_writeObject_Collections_SynchronizedCollection

The construct :

while (it1.hasNext()) {
    equals = equals && it1.next().equals(it2.next());
}

is broken because the RHS of the logical-and expression need not
be evaluated if equals == false.

The tests are still broken because they assume that iteration
order for HashSets is stable across serialization, but at least they
won't loop indefinitely if that's the case. These issues will be fixed
as part of the jdk8u60 upgrade.

Change-Id: I404b40d57c0ee13199fe0a8bd6d0e2c629d74a80