URLConnectionTest: explicitly close() all the TestSSLContexts
TestSSLContext.create() uses a SSLServerSocketFactory to construct
a SSLServerSocket. Not closing that SSLServerSocket frequently (not
always - seems to be more common under vogar) causes the CloseGuard
inherited from AbstractPlainSocketImpl to fail the test on my okhttp
branch.

While it's unclear to my why the Closeguard doesn't detect this (and
fail the test) on current AOSP master head, I don't see any reason
not to close() them.

Change-Id: Ifffeb576b25e07375917dcfc0fdec3c6a9da1be9