URLConnectionTest: explicitly close() all the TestSSLContexts
This fixes the test failing when run on vogar.

Without the close(), the test failed because:

 - TestSSLContext.serverSocket.getImpl() instanceof SocksSocketImpl
   which extends AbstractPlainSocketImpl.
 - That base class in turn uses CloseGuard to enforce (best effort)
   that close() is being called.

I originally observed this in vogar on a local branch with an
updated version of OkHttp (2.7.5). I have reproduced it on
AOSP master head but have not reproduced it under cts-tradefed.
Therefore I conclude that there is an additional bug that caused
this test to pass in cts-tradefed when it should have failed; I
have filed bug 29533469 for that.

Change-Id: Ifffeb576b25e07375917dcfc0fdec3c6a9da1be9