Let testImplAccept_detectsClosedState() fail quickly when it fails.
Previously this test relied on the test runner enforcing a reasonable
timeout, which sometimes was not the case (e.g. bug 28998288).

Note that this test should only fail on devices that exhibited
bug 27763633, which was fixed in May. The quick failure relies on
the test runner recognizing failures from background threads, which
I did not verify.

Tested:
  - Checked that the test still passes.
  - Did not empirically verify that it fails quickly on an old build.
    This is just a best effort.

Bug: 29030362
Change-Id: I0e25c1be52afaae207e7701bbccdfc681cb0a458