Revert some non-functional changes to source
While integrating OpenJDK some non-functional changes were made
to the source in order to improve consistency and formatting.
Unfortunately, these produce conflicts when applying upstream
patches making it more difficult to see the functional changes.

On balance the improvement in readability and consistency is not
worth the cost of the upgrade so this change reverts those
changes that caused conflicts to simplify the task of upgrading.

Most of the changes are localized and is easy to see that they
have no impact on functionality, the exceptions are:

CollationElementIterator - getOffset() method moved to the same
position as it had in 7u40.

Collator - move icuColl, constructors, getAvailableLocales(),
and clone() members to the same position as they had in 7u40.

SimpleDateFormat - move cases for STANDALONE options after the
non-standalone ones to match the order in 7u40.

Bug: 29352743
Change-Id: Ifc6b0d2455e6330ddc7433d8c39b8f0dcc18f8b3