Fix BigDecimal multiplication for -1 * Long.MIN_VALUE
At least since Gingerbread, BigDecimal incorrectly got the result
-1 << 63 instead of the correct value +1 << 63. The same error
occurred for the multiplication (-1 << n) * (-1 << (63-n)) as well
as scaled versions of it (e.g (Long.MIN_VALUE * 1E-5) * (-1E+4)).

This bug is related to the assymetry that -2^63 can be represented by
a long but +2^63 can not. BigDecimal internally uses the concept of
bitLength. That term refers to the number of bits, not including the
sign bit, needed to represent a signed value; it therefore exhibits
the same asymmetry. The logic used to determine whether the result of
a calculation will fit into a long or need to be upgraded to a
BigInteger doesn't correctly account for the asymmetry.

The bitLength of Long.MAX_VALUE and Long.MIN_VALUE is 63. It was
considered to change the definition of bitLength such that it only
depends on the absolute magnitude of a value (increasing
Long.MIN_VALUE's bitLength 64), but that would be inconsistent with
BitInteger.bitLength() and therefore require corresponding changes
elsewhere in BigDecimal. The risk of introducing bugs with such a
change seems too great to be worth the benefit of fixing this
particular bug.

Therefore this CL adds more specialized logic to recognize that a
multiplication result of 2^63 will not fit into a long. This is
similar to the localized fix for the division case done in commits
52200aac041b0f019ecc8b91da5f34971c7b3112 and
e0840cc531ef6c534797a80b5e229fef5fec8570.

Since test coverage for BigDecimal was pretty poor, this CL adds
a few more tests.

Test: BigDecimalTest still passes

Bug: 19185440
Change-Id: I39d68b3294e1c95e0cdd3284068801c06bf4f3f1