Fix DateFormatSymbols.equals()/.hashCode()
DateFormatSymbols was always calculating/loading zoneStrings in its
equals() and hashCode() methods, even when in most cases it wasn't
needed.

Additionally .hashCode() depended only on the values for the first
timezone in zoneNames, which is a very indicator of similarity (see the
order of field comparisons in equals() for a more useful approach).

I've changed hashCode to use months and weekdays for the hashCode() and
equals() to avoid loading the zoneStrings if neither object has loaded
them before.

This is in preparation of a future commit that should remove the use of
zoneStrings except in cases where user code explicitly interacts with
them (via getZoneStrings()/getZoneStrings()).

Note that despite this work, DateFormatSymbols is still not a good
candidate for use as a key in a Map, as it is mutable.

Bug: 20879084
Change-Id: I7f8adc48a21ec394cef5488dd9231807be4f0a4d