Add tests and benchmarks for SimpleDateFormat.
Add two tests related to how SimpleDateFormat interacts with
getZoneStrings/setZoneStrings in DateFormatSymbols. These tests simply
check existing behaviour in preparation of upcoming changes.

The SimpleDateFormatBenchmark has a main() method in addition to the
usual caliper tests to allow measuring some first-call times. Otherwise
the impact of calling a method on cold caches would be lost in the
caliper tests.

Bug: 20879084
Change-Id: I01581a31566df8fa1b60ccce0c8ddecb0230b39e