Delegate time zone names handling to ICU in SDF.
Delegate parsing and formatting time zone names to the relevant ICU code
(android.icu.text.TimeZoneFormat and related classes).

This fixes an issue where the cache of time zone names was time dependent.

The old code that uses DateFormatSymbols.getZoneStrings() for parsing
and formatting is still used  as a fallback, but it's only used if the
DateFormatSymbols.setZoneStrings() has been called explicitly.

Performance of this code is between ~60-220% of the old code, depending
on the exact use case (Measured with SimpleDateFormatBenchmark from
an earlier commit).

Test: vogar libcore/luni/src/test/java/libcore/java/text/SimpleDateFormatTest.java
Bug: 20879084
Change-Id: Id9f70acad60d97b9879df14b5b5de6e89b209de3