WIP: Enforce the framework's UncaughtExceptionHandler isn't replaceable.
Provide a first party (@hide) libcore API to set an UEH that will
still run if an UEH provided by an app returns cleanly, throws, or
calls System.exit(). The framework's UEH will not run if the app's
UEH kills the process more thoroughly through other means.

The new setGuardUncaughtExceptionHandler() API is a third kind of
UEH next to the existing (public) setUncaughtExceptionHandler()
and setExceptionHandler() libcore APIs.

This is for discussion only and not intended to be submitted
without substantial changes and testing. As is, this change will
make it impossible for apps to not die in the case of uncaught
exceptions. We probably don't want to do that since it will likely
cause app compatibility issues, hence if we wanted to submit this
change then we'd want to also change framework's UEH to split the
logging logic (GuardUncaughtExceptionHandler) from the dying logic
DefaultUncaughtExceptionHandler)

Bug: 29624607
Change-Id: Ic6bd2f5cc20c6aeee394397cc76b6160216ca1df