WIP: Ensure uncaught exceptions are logged if process doesn't die hard enough.
Ways an UncaughtException can "not die hard enough" are: return,
throw, or call System.exit(). Process.killProcess(), blocking the
thread or going into an infinite loop don't count.

Normally, the framework's RuntimeInit.UncaughtHandler will log the
exception and kill the process. For apps that do not replace this
framework's handler, this happens to avoid duplicate logging because
the way the process is killed (via Process.killProcess()) is not
covered by this mechanism; this coupling is hacky.

Bug: 29624607
Change-Id: I1fe50a6ec2e91778f37d239a3be732fe00e29815