javax.security.auth.Subject: remove code that is useless in Android
Classes SecurityManager, AccessControl, AccessControlContext,
SecurityDomainCombiner, AuthPermission are "dummy" in Android (classes are
there but methods do nothing, or just return null) and as a result
much of the code in Subject is pointless.

Removing such code.

Also, update to javadocs/comments.

Note, still needs some changes to serialization as to be fully updated from
jdk8u60.

Bug: 29631070

Change-Id: I67f576b2d45826921add2fb16ef893d6687d3b68